feat: restructure TeacherDashboard with component separation
