+++ frontend/src/components/Pages/Dashboard/TeacherDashboard/components/NotificationModal.jsx
@@ -1,0 +1,114 @@
+import { Bell } from 'lucide-react'
+
+// Format relative time for notifications (moved inline)
+const formatRelativeTime = (timestamp) => {
+  if (timestamp === "Just now") return timestamp
+
+  const now = new Date()
+  const date = new Date(timestamp)
+  const diffInSeconds = Math.floor((now - date) / 1000)
+
+  if (diffInSeconds < 60) return "Just now"
+  if (diffInSeconds < 3600) return `${Math.floor(diffInSeconds / 60)} minutes ago`
+  if (diffInSeconds < 86400) return `${Math.floor(diffInSeconds / 3600)} hours ago`
+  return `${Math.floor(diffInSeconds / 86400)} days ago`
+}
+
+const NotificationModal = ({
+  notifications,
+  unreadCount,
+  onClose,
+  onMarkAsRead,
+  onMarkAllAsRead,
+  onDeleteNotification
+}) => {
+  return (
+    <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50">
+      <div className="bg-white rounded-lg shadow-xl w-full max-w-2xl max-h-[80vh] flex flex-col">
+        {/* Modal Header */}
+        <div className="px-6 py-4 border-b border-gray-200 flex justify-between items-center">
+          <h2 className="text-xl font-semibold text-gray-800">All Notifications</h2>
+          <div className="flex items-center space-x-2">
+            {unreadCount > 0 && (
+              <button
+                onClick={onMarkAllAsRead}
+                className="text-sm text-indigo-600 hover:text-indigo-800 font-medium"
+              >
+                Mark all as read
+              </button>
+            )}
+            <button
+              onClick={onClose}
+              className="text-gray-400 hover:text-gray-600 text-2xl font-bold"
+            >
+              ×
+            </button>
+          </div>
+        </div>
+
+        {/* Modal Content */}
+        <div className="flex-1 overflow-y-auto p-6">
+          {notifications.length > 0 ? (
+            <div className="space-y-4">
+              {notifications.map((notification) => (
+                <div
+                  key={notification.id}
+                  className={`p-4 rounded-lg border ${
+                    !notification.read ? "bg-indigo-50 border-indigo-200" : "bg-gray-50 border-gray-200"
+                  } hover:shadow-md transition-shadow`}
+                >
+                  <div className="flex justify-between items-start">
+                    <div className="flex-1 min-w-0">
+                      <div className="flex items-center space-x-2 mb-2">
+                        <h3 className="text-sm font-medium text-gray-800">{notification.title}</h3>
+                        {!notification.read && (
+                          <span className="h-2 w-2 bg-indigo-500 rounded-full"></span>
+                        )}
+                      </div>
+                      <p className="text-sm text-gray-600 mb-2">{notification.message}</p>
+                      <p className="text-xs text-gray-400">{formatRelativeTime(notification.time)}</p>
+                    </div>
+                    <div className="flex items-center space-x-2 ml-4">
+                      {!notification.read && (
+                        <button
+                          onClick={() => onMarkAsRead(notification.id)}
+                          className="text-xs text-indigo-600 hover:text-indigo-800 font-medium px-2 py-1 rounded"
+                        >
+                          Mark as read
+                        </button>
+                      )}
+                      <button
+                        onClick={() => onDeleteNotification(notification.id)}
+                        className="text-xs text-red-600 hover:text-red-800 font-medium px-2 py-1 rounded hover:bg-red-50"
+                      >
+                        Delete
+                      </button>
+                    </div>
+                  </div>
+                </div>
+              ))}
+            </div>
+          ) : (
+            <div className="text-center py-12">
+              <Bell className="h-16 w-16 mx-auto text-gray-300 mb-4" />
+              <h3 className="text-lg font-medium text-gray-700 mb-2">No notifications</h3>
+              <p className="text-gray-500">You're all caught up! New notifications will appear here.</p>
+            </div>
+          )}
+        </div>
+
+        {/* Modal Footer */}
+        <div className="px-6 py-4 border-t border-gray-200 flex justify-end">
+          <button
+            onClick={onClose}
+            className="px-4 py-2 bg-gray-100 text-gray-700 rounded-lg hover:bg-gray-200 transition-colors"
+          >
+            Close
+          </button>
+        </div>
+      </div>
+    </div>
+  )
+}
+
+export default NotificationModal
